+++ src/components/breadCrumb/BreadCrumb.tsx
@@ -1,32 +1,57 @@
 import styled from 'styled-components';
 import { Arrow } from '../styleGuide/icon';
 import { Text } from '../styleGuide/text/Text';
-import { useLocation } from 'react-router-dom';
-import { marginCssType } from '../../utils/distance';
+import { useLocation, Link } from 'react-router-dom';
 
-interface PropsType extends marginCssType {
+interface PropsType {
   className?: string;
-  pathToKorean: { [key: string]: string };
+  pathToKorean: any;
+  top?: number;
+  left?: number;
 }
 
-// 동적 라우팅
-// 링크 이동
-export const BreadCrumb = ({ className, pathToKorean, margin }: PropsType) => {
+export const BreadCrumb = ({
+  className,
+  pathToKorean,
+  top = 86,
+  left = 330,
+}: PropsType) => {
+  const path: string[] = useLocation().pathname.split('/').slice(1);
+  const pathList = [];
+  const LinkList = [`/${path[0]}`];
+  for (let i = 0; i < path.length; i++) {
+    if (pathToKorean[path[i]]) {
+      if (pathToKorean[path[i]].index)
+        pathList.push(pathToKorean[path[i]].index);
+      else if (typeof pathToKorean[path[i]] === 'string')
+        pathList.push(pathToKorean[path[i]]);
+      pathToKorean = pathToKorean[path[i]];
+    } else if (pathToKorean['dynamic']) {
+      pathList.push(pathToKorean['dynamic']);
+      pathToKorean = pathToKorean['dynamic'];
+      LinkList[LinkList.length - 2] = `${LinkList[LinkList.length - 2]}/${
+        path[i]
+      }`;
+    }
+
+    if (i > 0) LinkList.push(`${LinkList[LinkList.length - 1]}/${path[i]}`);
+  }
   return (
-    <_Wrapper margin={margin} className={className}>
-      {/* {useLocation()
-        .pathname */}
-      {'/notice/my-page'
-        .split('/')
-        .slice(1)
-        .map((item, idx, arr) => (
+    <_Wrapper left={left} top={top} className={className}>
+      {pathList &&
+        pathList.map((item, idx, arr) => (
           <>
-            <Text
-              size="bodyS"
-              color={arr.length !== idx + 1 ? 'gray5' : 'gray10'}
-            >
-              {pathToKorean[item]}
-            </Text>
+            {arr.length !== idx + 1 ? (
+              <Link to={LinkList[idx]}>
+                <Text size="bodyS" color={'gray5'}>
+                  {item}
+                </Text>
+              </Link>
+            ) : (
+              <Text size="bodyS" color={'gray10'} cursor="pointer">
+                {item}
+              </Text>
+            )}
             {arr.length !== idx + 1 && (
               <Arrow colorKey="gray5" direction="right" />
             )}
@@ -36,8 +61,11 @@ export const BreadCrumb = ({ className, pathToKorean, margin }: PropsType) => {
   );
 };
 
-const _Wrapper = styled.div<marginCssType>`
+const _Wrapper = styled.div<{ top: number; left: number }>`
+  position: absolute;
   display: flex;
   align-items: center;
   gap: 12px;
+  top: ${({ top }) => top}px;
+  left: ${({ left }) => left}px;
 `;