+++ src/components/input/index.tsx
@@ -13,6 +13,7 @@ interface PropsType extends marginCssType, errorMsgPropsType, labelPropsType {
   name: string;
   className?: string;
   value: string | number;
+  limit?: number;
 }
 
 /** input 에러 상태일 때에는 포커싱이 파란색..? */
@@ -27,18 +28,31 @@ export const Input = ({
   value,
   className,
   margin,
+  limit,
 }: PropsType) => {
   const [isOpen, setIsOpen] = useState<boolean>(false);
+  const [isInputClicked, setIsInputClicked] = useState<boolean>(false);
   return (
     <_Wrapper margin={margin} className={className}>
       <Label label={label} />
       <_InputWrapper width={width}>
         <_Input
+          onWheel={(e) => {
+            //@ts-ignore
+            document.activeElement.blur();
+          }}
+          onFocus={() => {
+            setIsInputClicked(true);
+          }}
+          onBlur={() => {
+            setIsInputClicked(false);
+          }}
           onChange={onChange}
           type={(isOpen && 'text') || type}
           errorMsg={errorMsg}
-          placeholder={placeholder}
+          placeholder={isInputClicked ? '' : placeholder}
           value={value}
+          maxLength={limit}
           name={name}
         />
         {type == 'password' && (